move worker.py to jagereye as a library, and let app developer can specify msg_bus server
