+++ jagereye/worker/__init__.py
@@ -1,0 +1,3 @@
+from jagereye.worker.worker import Worker
+
+__all__ = ['Worker']
